requirements.txt
chromadb
langchain
langchain_text_splitters
ollama
pydantic
pymupdf
python-dotenv
requests